requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==38.0.4
decorator==4.4.2
Deprecated==1.2.13
Django==4.1.4
django-cors-headers==3.12.0
django-filter==22.1
django-oauth-toolkit==1.7.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
gunicorn==20.1.0
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.7
jwcrypto==1.4.2
moviepy==1.0.3
numpy==1.24.0
oauthlib==3.2.2
phonenumbers==8.12.48
Pillow==9.3.0
proglog==0.1.10
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.13
wrapt==1.14.1
